A bunch more progress mainly involving the tutorial
Added many many sprite assets to the game like the sign messages and the background tiles for the tutorial.

Added functionality to the signs so when you approach and press "E" they display and when you walk away they close.

Got more of the tutorial done, need to make weapons for the weapons sections though.

Switched back from C# PlayerController to Visual Scripting PlayerController since the C# one is still not working as intended.